refactor(Product): Jpa로 변경
